add sqlc

+++ internal/dto/list_response.go
@@ -1,6 +1,6 @@
 package dto
 
 type ListResponse[T any] struct {
-	List  T   `json:"list,omitempty"`
-	Total int `json:"total,omitempty"`
+	List  T     `json:"list,omitempty"`
+	Total int64 `json:"total,omitempty"`
 }
